requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.10.3
bcc==0.29.1
blessed==1.20.0
blinker==1.7.0
Bottleneck==1.5.0
Brlapi==0.8.5
cachetools==5.4.0
catboost==1.2.3
certifi==2024.7.4
cffi==1.17.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.6
cloud-init==25.1.4
colorama==0.4.6
command-not-found==0.3
commentjson==0.9.0
configobj==5.0.8
contourpy==1.2.1
cryptography==42.0.2
cupshelpers==1.0
cycler==0.12.1
dataclasses-json==0.6.7
dbus-python==1.3.2
defer==1.0.6
distro==1.9.0
distro-info==1.7+build1
Django==4.1
django-cors-headers==3.13.0
django-dump-die==0.1.7
django-picklefield==3.2
django-q==1.3.9
docstring_parser==0.16
ElasticEmail @ git+https://github.com/elasticemail/elasticemail-python.git@533ab5a06e3bac64138db6f3ecf61676730ff604
emoji==2.10.1
et_xmlfile==2.0.0
faiss-cpu==1.11.0
feature-engine==1.8.0
fonttools==4.53.1
frozenlist==1.4.1
google-api-core==2.19.0
google-auth==2.30.0
google-cloud-aiplatform==1.48.0
google-cloud-bigquery==3.29.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-genai==1.24.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
graphviz==0.20.3
greenlet==3.0.3
grpc-google-iam-v1==0.14.0
grpcio==1.70.0
grpcio-status==1.62.3
gunicorn==23.0.0
gyp==0.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
httpx-sse==0.4.1
idna==3.7
Jinja2==3.1.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.67
langchain-google-vertexai==2.0.27
langchain-ollama==0.3.3
langchain-text-splitters==0.3.8
langsmith==0.4.4
language-selector==0.1
lark-parser==0.7.8
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.2
mdurl==0.1.2
multidict==6.0.5
mypy_extensions==1.1.0
netaddr==0.8.0
netifaces==0.11.0
numexpr==2.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
ollama==0.5.1
openpyxl==3.1.5
orjson==3.10.18
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pexpect==4.9.0
pillow==10.4.0
plotly==5.23.0
proto-plus==1.24.0
protobuf==4.25.4
psutil==6.1.0
ptyprocess==0.7.0
pyarrow==19.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.25.1
pycparser==2.22
pycryptodome==3.20.0
pycups==2.0.1
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
Pygments==2.17.2
PyGObject==3.48.2
PyJWT==2.4.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyrsistent==0.20.0
pyserial==3.5
python-apt==2.7.7+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.49+ubuntu2
python-dotenv==1.0.1
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.2
redis==3.5.3
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.22.3
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
setuptools==75.0.0
shapely==2.0.7
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.39
sqlparse==0.5.1
statsmodels==0.14.2
systemd-python==235
tenacity==9.0.0
threadpoolctl==3.5.0
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
Unidecode==1.3.8
urllib3==2.2.2
validators==0.35.0
vertexai==0.0.1
wadllib==1.3.6
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.0
wheel==0.42.0
xdg==5
xkit==0.0.0
yarl==1.9.4
zstandard==0.23.0
